requirements.txt
python-telegram-bot
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv 